requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==2.1.0
